refactor libs and update logger call signature
