+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@kksiuda/http-session",
-  "version": "0.9.31",
+  "version": "0.9.32",
   "description": "Http session and cookies management",
   "main": "./dist/index.js",
   "author": "Krzysztof Siuda",
