+++ src/lib/noOpLogger.ts
@@ -1,0 +1,11 @@
+function noOp() {
+  //
+}
+
+export const noOpLogger = {
+  debug: noOp,
+  info: noOp,
+  warn: noOp,
+  error: noOp,
+  namespace: () => noOpLogger,
+};
